Move ThreadContextElement to common
