+++ kotlinx-coroutines-core/jsAndWasmShared/src/internal/ThreadLocal.kt
@@ -5,6 +5,7 @@ internal actual class CommonThreadLocal<T> {
     @Suppress("UNCHECKED_CAST")
     actual fun get(): T = value as T
     actual fun set(value: T) { this.value = value }
+    actual fun remove() { value = null }
 }
 
 internal actual fun<T> commonThreadLocal(name: Symbol): CommonThreadLocal<T> = CommonThreadLocal()
